Implements `doTransactionDetailsMatchPurchaseInfo:withPurchaseInfo:`.
This version differs from Apple's implementation in that it checks for all optional cases except `requestData`.
The reason `requestData` is avoided is because this is a field Apple reserve for future uses and thus its behaviour is subject to change. Guarding against it will make the code less future proof.
A detail explanation is available as inline comments.