[CRITICAL BUG] Fixed restoration bug caused by incorrectly check for unique identifier for vendor.
Every (re-)installation generates a new unique identifier for vendor.
Every purchase and restoration receipt will be tag with this new unique identifier.
But the latest receipt info might have a unique identifier for vendor from another device, from a previous installation, etc.
So we should only check if the purchase info matches receipt for device for receipt we restored with this device.
We should not check for unique identifier for vendor in latest receipt info.
This has caused a critical bug where restoration will always fails.

This commit attempts to fixed this issue.